+++ packages/plugin-packages/mashroom-portal-remote-app-registry/src/js/jobs/mashroom-bootstrap-background-job.ts
@@ -4,9 +4,9 @@ import RegisterPortalRemoteAppsBackgroundJob from './RegisterPortalRemoteAppsBac
 import type {MashroomBackgroundJobPluginBootstrapFunction} from '@mashroom/mashroom-background-jobs/type-definitions';
 
 const bootstrap: MashroomBackgroundJobPluginBootstrapFunction = (pluginName, pluginConfig, pluginContextHolder) => {
-    const { socketTimeoutSec, registrationRefreshIntervalSec } = pluginConfig;
+    const { socketTimeoutSec, registrationRefreshIntervalSec, unregisterAppsAfterScanErrors } = pluginConfig;
 
-    const registerBackgroundJob = new RegisterPortalRemoteAppsBackgroundJob(socketTimeoutSec, registrationRefreshIntervalSec, pluginContextHolder);
+    const registerBackgroundJob = new RegisterPortalRemoteAppsBackgroundJob(socketTimeoutSec, registrationRefreshIntervalSec, unregisterAppsAfterScanErrors, pluginContextHolder);
     context.backgroundJob = registerBackgroundJob;
 
     // Run immediately